+++ src/User/Auth0/Auth0ManagementTokenRepository.php
@@ -4,9 +4,11 @@ declare(strict_types=1);
 
 namespace CultuurNet\UDB3\User\Auth0;
 
+use CultuurNet\UDB3\User\ManagementToken;
+
 interface Auth0ManagementTokenRepository
 {
-    public function token(): ?Auth0Token;
+    public function token(): ?ManagementToken;
 
-    public function store(Auth0Token $token): void;
+    public function store(ManagementToken $token): void;
 }
