+++ src/User/Auth0/InMemoryRepository.php
@@ -4,16 +4,18 @@ declare(strict_types=1);
 
 namespace CultuurNet\UDB3\User\Auth0;
 
+use CultuurNet\UDB3\User\ManagementToken;
+
 class InMemoryRepository implements Auth0ManagementTokenRepository
 {
-    private ?Auth0Token $token = null;
+    private ?ManagementToken $token = null;
 
-    public function token(): ?Auth0Token
+    public function token(): ?ManagementToken
     {
         return $this->token;
     }
 
-    public function store(Auth0Token $token): void
+    public function store(ManagementToken $token): void
     {
         $this->token = $token;
     }
